move NavCard and Generator to welcome forder
