+++ apps/ideaspace/src/components/modules/WelcomePage/IdeaGeneratorButton/IdeaGeneratorButton.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 
-import Modal from '../../common/Modal';
+import Modal from '../../../common/Modal';
 
 import { ModalWrapper } from './StyledIdeaGeneratorButton';
