+++ apps/ideaspace/src/components/modules/WelcomePage/IdeaGeneratorButton/StyledIdeaGeneratorButton.js
