+++ packages/UI/src/components/organisms/Footer/RandomQuote/RandomQuote.js
@@ -1,3 +1,4 @@
+import React, { useState } from 'react';
 import Wrapper from './StyledRandomQuote';
 
 // Retrieve all quotes from our content layer
@@ -6,7 +7,11 @@ const quotes =
 
 export default function RandomQuote() {
   // Select and display a random quote from our quotes collection
-  const randomQuote = quotes[parseInt(quotes.length * Math.random(), 10)];
+  const [randomNum, setRandomNum] = useState(0);
+  React.useEffect(() => {
+    setRandomNum(parseInt(quotes.length * Math.random(), 10));
+  }, []);
+  const randomQuote = quotes[randomNum];
   return (
     <Wrapper>
       &quot;{randomQuote.body}&quot; - {randomQuote.author}
