+++ packages/UI/src/components/organisms/OpenResponse/OpenResponse.tsx
@@ -39,7 +39,7 @@ function OpenResponse({
             autoCorrect="off"
             autoComplete="off"
             valid={touched && !error}
-            error={error}
+            error={touched ? error : ''}
             disabled={disabled}
             onChange={onChange}
             value={value}
